:green_heart: chore: ABOUTME 섹션 폴더 분리 #26
